Adds Email to Appoinment Form
The request appointment form now uses a php script to send an email
(right now it is my email for testing purposes).  The script runs
when running the app on a php server.  The only problem is the email
does not send.  This might be something to test once the site is live.

You can test by starting a server `php -S localhost:8000`